pyproject.toml
[project]
name = "cellsensei"
version = "0.1.0"
description = "Automated feedback and testing for Jupyter notebooks."
authors = [
    { name = "Michael Borck", email = "michael@borck.dev" },
]
dependencies = [
    "python-fasthtml",
    "uvicorn[standard]",
    "nbformat",
    # Celery and Redis/broker will be added later
]
requires-python = ">=3.9" # Choose your minimum Python version

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
    "pytest",
    "uv", # If you want to ensure uv is available for dev tasks
    "mkdocs", # For documentation later
    "mkdocs-material", # Theme for mkdocs
    # "celery[redis]", # Add when you reach that milestone
]

[tool.ruff]
# Ruff configuration options go here
# e.g., line-length = 88
# select = ["E", "F", "W", "I", "UP", "B", "C4"] # Common selections

[tool.ruff.format]
# Ruff formatter options

[tool.mypy]
# Mypy configuration options

[tool.pytest.ini_options]
# Pytest configuration
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests", # Assuming you have a top-level tests directory for app tests
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

# If you decide to use Hatch as your build backend (popular with pyproject.toml)
# [build-system]
# requires = ["hatchling"]
# build-backend = "hatchling.build"
# [tool.hatch.version]
# path = "src/cellsensei/__about__.py" # Example if you put code in src/
